Update Kalman_Filter_class_development.ipynb
Removed LinearGaussianModel and NonlinearGaussianModel. System dynamics and models are specific to problems and it will be better if it was defined on a case-by-case basis. It is quite frankly almost impossible to have a universal system model for all the various types of problems. So, now the workflow is as such:
1) Define states and initial conditions for state vector and covariances
2) Define ur model - create a Class that defines the system dynamics, F, Q, H, R, f, h, F_jacobian, H_jacobian
3) Run the Kalman Filter
4) Analyze the data

So now the logic to check and validate matrices defined in the models are all internalized to the variants of the kalman filter classes. While this may increase an effort while writing newer variants of kalman, it will be worth the effort still.